Add BC7 support and D3D12 resource fixes
Add BC7 codec and integrate BC7 into the D3D12 renderer and tools, improve sun CB handling, and make D3D12 resource lifetime safer.

Summary of changes:
- Add ImageCodecBC7 (encode/decode) and vendor bc7enc/bc7decomp third-party sources; update DisplayD3D12 project to compile/link them and include new headers.
- Register ColorFormat::BC7 in Image and map DXGI_FORMAT_BC7 ↔ ColorFormat::BC7 in D3D12 code paths and PrimitiveSurface helpers.
- Extend CBPerFrame with vSunDir/vSunColor and vSunWorldPos and expose/update them in bindPerFrameCB for shaders; make DisplayDevice sun-candidate logic sticky across frames and reset only the per-frame comparator.
- Add Material::flushSurfaceCache() call during device release to avoid stale/zombie surfaces when tearing down/recreating devices.
- Introduce DisplayDeviceD3D12::safeDeferReleaseResource to handle dangling device pointers (MDI/teardown races) and replace many primitive release paths to use it; add aligned new/delete to ensure 16-byte allocation for the class.
- Improve ImageCodecDXT to slab Compress/Decompress calls and report progress (Progress::report) so large mips don't hang the UI.
- Optimize InStream::readHexLine to read large chunks instead of one-byte syscalls; include Progress header where needed.

Rationale: enable high-quality BC7 compressed textures for DX12, reduce teardown-related resource/heap errors, expose usable sun data to shaders without frame-order artifacts, and improve encoding/IO responsiveness and robustness.

+++ Tools/ResourcerDoc/PropertyList.cpp
@@ -624,10 +624,14 @@ BOOL CPropertyList::PreCreateWindow(CREATESTRUCT& cs)
 	return TRUE;
 }
 
-void CPropertyList::MeasureItem(LPMEASUREITEMSTRUCT lpMeasureItemStruct) 
+void CPropertyList::MeasureItem(LPMEASUREITEMSTRUCT lpMeasureItemStruct)
 {
-	int nIndex = lpMeasureItemStruct->itemID;
-	if ( nIndex < GetCount() )
+	// itemID is UINT and may be (UINT)-1 during the one-shot WM_MEASUREITEM
+	// LBS_OWNERDRAWFIXED sends at list-box creation (no items yet, value
+	// undefined per Win32 docs).  Compare unsigned so 0xFFFFFFFF < 0 is
+	// correctly false and we fall through to the default height.
+	UINT nIndex = lpMeasureItemStruct->itemID;
+	if ( nIndex < (UINT)GetCount() )
 	{
 		PropertyListItem * pItem = (PropertyListItem *)GetItemDataPtr( nIndex );
 		lpMeasureItemStruct->itemHeight = pItem->nHeight; //pixels
